[macOS] Date/time inputs should focus the next editable component when entering a separator key https://bugs.webkit.org/show_bug.cgi?id=216661 <rdar://problem/69091856>
Reviewed by Wenson Hsieh.

Source/WebCore:

NSDatePicker focuses the next editable field when entering a separator
key. We should match this behavior to be consistent with the rest of
the system.

* html/shadow/DateTimeFieldElement.cpp:
(WebCore::DateTimeFieldElement::defaultKeyboardEventHandler):

Tools:

* WebKitTestRunner/mac/EventSenderProxy.mm:
(WTR::EventSenderProxy::keyDown):

Add support for colon, semicolon, and comma keys. The keycode values are
taken from the virtual keycode values in HIToolbox/Events.h.

Note that the keycode for colon and semicolon is the same, as there is
no distinct value for colon. This is safe as it matches the Javascript
behavior.

LayoutTests:

Added new test cases to the existing set of keyboard tests for date/time inputs.

* fast/forms/date/date-editable-components/date-editable-components-keyboard-events-expected.txt:
* fast/forms/date/date-editable-components/date-editable-components-keyboard-events.html:
* fast/forms/datetimelocal/datetimelocal-editable-components/datetimelocal-editable-components-keyboard-events-expected.txt:
* fast/forms/datetimelocal/datetimelocal-editable-components/datetimelocal-editable-components-keyboard-events.html:
* fast/forms/time/time-editable-components/time-editable-components-keyboard-events-expected.txt:
* fast/forms/time/time-editable-components/time-editable-components-keyboard-events.html:


Identifier: 229519@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@267281 268f45cc-cd09-0410-ab3c-d52691b4dbfc