[LFC][IFC] Atomic inline-level box with margin is mispositioned https://bugs.webkit.org/show_bug.cgi?id=216842
Reviewed by Antti Koivisto.

The inline box that an atomic inline-level box generates has the height of the margin box. Therefore the inline box's logical top position
is the position of the top edge of the margin box.
When converting the inline box geometry back to the layout box, we need to offset the top position (border box top edge)
with the value of the used margin before.

* layout/inlineformatting/InlineFormattingContext.cpp:
(WebCore::Layout::InlineFormattingContext::computeGeometryForLineContent):


Identifier: 229635@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@267437 268f45cc-cd09-0410-ab3c-d52691b4dbfc