[css-grid] Clear the override width for computing percent margins https://bugs.webkit.org/show_bug.cgi?id=209461
Reviewed by Manuel Rego Casasnovas.

LayoutTests/imported/w3c:

Update test expectations.

The 2 grid-items-percentage-paddings tests are now expected to have some
failures because of bug 191964. In fact they were already failing before
this patch, as can be seen in wpt.fyi:
https://wpt.fyi/results/css/css-grid/grid-items?q=grid-items-percentage-paddings-vertical-%20-002.html
It's just that the bug that is being fixed made the test runner think
that the tests were passing, probably some racy condition or such.

* web-platform-tests/css/css-grid/grid-definition/grid-minimum-contribution-with-percentages-expected.txt:
* web-platform-tests/css/css-grid/grid-items/grid-item-dynamic-min-contribution-001-expected.txt:
* web-platform-tests/css/css-grid/grid-items/grid-items-minimum-height-orthogonal-001-expected.txt:
* web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-vertical-lr-002-expected.txt:
* web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-vertical-rl-002-expected.txt:

Source/WebCore:

When calculating the min-content contribution of a grid item of an auto
sized grid track we must consider the grid item's margin. When the grid
item's area is indefinite, a percent margin is resolved to zero.
However, when performing a relayout, the percent margin may be solved
against the previously computed grid area, since the grid item has
already an OverrideContainingBlockLogicalWidth value.

In order to re-compute the percent margin properly, we need to clear
the previously override value. It's important to be careful of not
clearing the override value set during intrinsic size, since we need
it for the actual layout phase. Hence, we only reset the 'override'
value when we are executing a definite strategy.

This is a reland of r261841, which was reverted because it broke spotify.
This problem is now avoided thanks to bug 210089.

Tests: imported/w3c/web-platform-tests/css/css-grid/grid-definition/grid-minimum-contribution-with-percentages.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-item-dynamic-min-contribution-001.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-minimum-height-orthogonal-001.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-margins-003.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-margins-004.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-margins-005.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-margins-006.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-margins-007.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-margins-008.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-margins-009.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-margins-010.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-005.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-006.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-009.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-010.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-vertical-lr-002.html
       imported/w3c/web-platform-tests/css/css-grid/grid-items/grid-items-percentage-paddings-vertical-rl-002.html

* rendering/GridTrackSizingAlgorithm.cpp:
(WebCore::hasRelativeMarginOrPaddingForChild):
(WebCore::hasRelativeOrIntrinsicSizeForChild):
(WebCore::shouldClearOverrideContainingBlockContentSizeForChild):
(WebCore::GridTrackSizingAlgorithmStrategy::minSizeForChild const):
(WebCore::GridTrackSizingAlgorithmStrategy::minLogicalSizeForChild const):
(WebCore::DefiniteSizeStrategy::minLogicalSizeForChild const):
(WebCore::DefiniteSizeStrategy::minContentForChild const):
* rendering/GridTrackSizingAlgorithm.h:

LayoutTests:

Update test expectations.

* TestExpectations:


Identifier: 229687@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@267503 268f45cc-cd09-0410-ab3c-d52691b4dbfc