CSS serialization expects comments between certain tokens https://bugs.webkit.org/show_bug.cgi?id=216947
Reviewed by Darin Adler.

LayoutTests/imported/w3c:

The serialize-consecutive-tokens test now passes as expected. We fail the anplusb-selector-parsing
test because we consider `n-b` to be a DimensionToken followed by a NumberToken but this is incorrect.
I'll fix that test in a follow up patch.

* web-platform-tests/css/css-syntax/serialize-consecutive-tokens-expected.txt:
* web-platform-tests/css/selectors/anplusb-selector-parsing-expected.txt:

Source/WebCore:

This patch brings us up to spec on inserting comments while
serializing CSS tokens. The spec says that we should always be
able to round trip a serialized output. Weirdly, for reasons I
don't understand, the spec requires a comment rather than
whitespace so that's what this patch does. Since there are 32
token types building a table for every combination would have been
massive. Instead, this patch uses a constexpr function to generate
the table for the set of next (non-delimiter) tokens that require
a separating comment. This makes it a bit more readable while,
hopefully, not sacrificing too much performance. For delimiters
requiring separating comments, we check for those in a helper
lambda after checking the token table.

* css/parser/CSSParserToken.cpp:
(WebCore::buildNextTokenNeedsCommentTable):
(WebCore::CSSParserToken::serialize const):
* css/parser/CSSParserToken.h:
* css/parser/CSSParserTokenRange.cpp:
(WebCore::CSSParserTokenRange::serialize const):

LayoutTests:

Rebasline CSS tests for better serialization. The css-selectorText
test produces different output since we tokenize the <an-plus-b>
syntax incorrectly. I don't know if this means we evaluate it
incorrectly, however. In a follow up patch I will fix the the
tokenization issue so the comments will no longer appear.

* fast/dom/css-selectorText-expected.txt:


Identifier: 229888@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@267766 268f45cc-cd09-0410-ab3c-d52691b4dbfc