[GPU Process] RemoteImageBufferProxy should ensure the state stack of the context is cleared before destruction https://bugs.webkit.org/show_bug.cgi?id=208876
Reviewed by Tim Horton.

When running canvas layout tests in debug WebKit, the assertion in `~GraphicsContext()` which verifies that
`m_stack.isEmpty()` fires. This occurs because the matching call to restore may still be buffered in the web
process when the remote image buffer proxy in the GPU process is destroyed; however, this may also occur if the
web process terminates without having sent the matching calls to restore the graphics context.

We can fix both of these scenarios by manually unwinding the context's state stack when destroying the
`RemoteImageBufferProxy`. Original patch by Said Abou-Hallawa.

* GPUProcess/graphics/RemoteImageBufferProxy.h:
(WebKit::RemoteImageBufferProxy::~RemoteImageBufferProxy):


Identifier: 229893@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@267772 268f45cc-cd09-0410-ab3c-d52691b4dbfc