[Media in GPU Process] Web process is not allowed to set the AudioSession category https://bugs.webkit.org/show_bug.cgi?id=217236
Reviewed by Eric Carlson.

We disabled setting the audio session category in the Web process when media is playing
in the GPU Process (r257367). But since RemoteAudioSession and RemoteAudioSessionProxy
are working now in the Web process and GPU process, we need to allow setting the audio
session category in the Web process.

This patch also includes an optimization to reduce the IPC traffic of the audio session
category changes.

* GPUProcess/media/RemoteAudioSessionProxy.cpp:
(WebKit::RemoteAudioSessionProxy::setCategory):
* GPUProcess/media/RemoteAudioSessionProxyManager.cpp:
(WebKit::RemoteAudioSessionProxyManager::setCategoryForProcess):
We need to check the category and policy in RemoteAudioSessionProxy instead of
RemoteAudioSessionProxyManager. Otherwise, the change will be ignored.

* WebProcess/GPU/media/RemoteAudioSession.cpp:
(WebKit::RemoteAudioSession::setCategory):
Only send RemoteAudioSessionProxy::SetCategory message if the category or the policy
is changed.
* WebProcess/GPU/media/RemoteAudioSession.h:

* WebProcess/WebPage/WebPage.cpp:
(WebKit::m_limitsNavigationsToAppBoundDomains):
Revert the change in r257367.


Identifier: 230040@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@267929 268f45cc-cd09-0410-ab3c-d52691b4dbfc