REGRESSION(r267399): [WPE] Build broken with ENABLE_ACCESSIBILITY=OFF https://bugs.webkit.org/show_bug.cgi?id=217380
Reviewed by Philippe Normand.

Before the ATK::Bridge target got introduced in r267399 a couple of CMake variables were
used, which were empty when the accessility support was disabled and nothing was being
added to the WebKit and WPEToolingBackends targets. Now that an imported target is used
it is needed to add it as a dependency conditionally to avoid using an undefined target
name.

Source/WebKit:

No new tests needed.

* PlatformWPE.cmake: Use the ATK::Bridge target only when ENABLE_ACCESSIBILITY is set.

Tools:

* wpe/backends/CMakeLists.txt: Use the ATK::Bridge target only when ENABLE_ACCESSIBILITY is set.


Identifier: 230145@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268046 268f45cc-cd09-0410-ab3c-d52691b4dbfc