REGRESSION: Date/time pickers are not displayed in UIWebViews https://bugs.webkit.org/show_bug.cgi?id=217341 <rdar://problem/66881739>
Reviewed by Wenson Hsieh.

Source/WTF:

Date/time inputs were incorrectly disabled by default on
PLATFORM(IOS_FAMILY) in WebPreferencesInternal.

* Scripts/Preferences/WebPreferencesInternal.yaml:

Tools:

Added regression tests to ensure this preference is not accidently
modified in the future.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKitLegacy/ios/DateTimeInputsAccessoryViewTests.mm: Added.
(-[DateTimeInputsTestsLoadingDelegate webViewDidFinishLoad:]):
(runTestWithInputType):

Note that [UIDatePicker setDatePickerMode:] had to be swizzled to an
empty implementation as the UIKit version on the bots attempts to
create a QoS interval upon setting the initially selected row in
a UITableView when the date picker is presented. This fails because
TestWebKitAPI isn't actually an app. To avoid an exception, the method
is swizzled out.


Identifier: 230192@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268123 268f45cc-cd09-0410-ab3c-d52691b4dbfc