REGRESSION(r268123): [ iOS ] 4 TestWebKitAPI.WebKitLegacy failures: File name unavailable. Unknown C++ exception thrown in the test body. https://bugs.webkit.org/show_bug.cgi?id=217445 <rdar://problem/70063370>
Reviewed by Wenson Hsieh.

The patch which introduced these tests swizzled [UIDatePicker setDatePickerMode:]
to avoid a crash within UIKit, resulting from the fact that TestWebKitAPI isn't
actually an app.

However, this was not a reliable solution since UIKit was still attempting to
create a QoS interval. To fix, swizzle out a method much closer to the source
of the exception: [UITableView cellForRowAtIndexPath:].

* TestWebKitAPI/Tests/WebKitLegacy/ios/DateTimeInputsAccessoryViewTests.mm:
(cellForRowAtIndexPath):
(runTestWithInputType):


Identifier: 230232@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@268169 268f45cc-cd09-0410-ab3c-d52691b4dbfc